feat: implement TenantsPage component with tenant management functionality and integrate into dashboard
