+++ src/components/common/page-header.tsx
@@ -5,11 +5,15 @@ interface PageHeaderProps {
 
 export const PageHeader = ({ title, subtitle }: PageHeaderProps) => {
 	return (
-		<div className="flex items-center justify-between">
+		<div className="flex flex-col items-start justify-between">
 			<div>
-				<h1 className="text-3xl font-bold tracking-tight">{title}</h1>
+				<h1 className="text-3xl font-bold">{title}</h1>
 			</div>
-			{subtitle && <div className="flex items-center gap-2">{subtitle}</div>}
+			{subtitle && (
+				<div className="text-muted-foreground flex items-center">
+					{subtitle}
+				</div>
+			)}
 		</div>
 	);
 };
