+++ src/components/tenants/index.ts
@@ -1,0 +1,1 @@
+export * from "@/components/tenants/tenants-page";
